[NCLSUP-1343] Skip deleting temp indy repo for some build status
Those build statuses are:
- NO_REBUILD_REQUIRED
- REJECTED_*

It's just a tiny optimization since in those status, no Indy and
Causeway repos should have been created anyways.

+++ build-coordinator/src/main/java/org/jboss/pnc/coordinator/maintenance/TemporaryBuildsCleaner.java
@@ -18,6 +18,7 @@
 package org.jboss.pnc.coordinator.maintenance;
 
 import org.jboss.pnc.enums.ArtifactQuality;
+import org.jboss.pnc.enums.BuildStatus;
 import org.jboss.pnc.enums.ResultStatus;
 import org.jboss.pnc.mapper.api.BuildMapper;
 import org.jboss.pnc.model.Artifact;
@@ -110,10 +111,17 @@ public class TemporaryBuildsCleaner {
                         + buildRecord.getBuiltArtifacts() + "; Dependencies: " + buildRecord.getDependencies());
 
         String externalBuildId = BuildMapper.idMapper.toDto(buildRecord.getId());
-        Result result = remoteBuildsCleaner.deleteRemoteBuilds(buildRecord);
-        if (!result.isSuccess()) {
-            log.error("Failed to delete remote temporary builds for BR.id:{}.", buildRecord.getId());
-            return new Result(externalBuildId, ResultStatus.FAILED, "Failed to delete remote temporary builds.");
+
+        BuildStatus buildStatus = buildRecord.getStatus();
+        if (buildStatus.isFinal() && buildStatus != BuildStatus.NO_REBUILD_REQUIRED
+                && !buildStatus.name().startsWith("REJECTED")) {
+            // there's only an indy repository if the build actually tried to run. NO_REBUILD_REQUIRED and REJECTED_*
+            // never ran
+            Result result = remoteBuildsCleaner.deleteRemoteBuilds(buildRecord);
+            if (!result.isSuccess()) {
+                log.error("Failed to delete remote temporary builds for BR.id:{}.", buildRecord.getId());
+                return new Result(externalBuildId, ResultStatus.FAILED, "Failed to delete remote temporary builds.");
+            }
         }
 
         removeBuiltArtifacts(buildRecord);
